Backend/ML-workspace/requirements.txt
pdfplumber>=0.11.0
google-generativeai>=0.8.0
pydantic>=2.9.0
python-dotenv>=1.0.0
